requirements.txt
beautifulsoup4==4.13.4
datasets==3.6.0
docx2txt==0.9
fitz==0.0.1.dev2
gradio==5.35.0
jieba==0.42.1
langchain==0.3.26
langchain_community==0.3.27
langchain_core==0.3.68
langchain_huggingface==0.3.0
langchain_text_splitters==0.3.8
langdetect==1.0.9
Markdown==3.8.2
matplotlib==3.10.3
modelscope==1.27.0
numpy==2.3.1
onnxruntime_gpu==1.20.1
optimum==1.26.1
pandas==2.3.0
peft==0.16.0
Pillow==11.3.0
PyPDF2==3.0.1
PyYAML==6.0.2
PyYAML==6.0.2
Requests==2.32.4
rouge==1.0.1
scipy==1.16.0
sentence_transformers==4.1.0
swanlab==0.6.4
torch==2.2.1+cu118
tqdm==4.67.1
transformers==4.51.0
